requirements.txt
streamlit
onnxruntime
numpy
pillow
opencv-python-headless